refactor(palette): flatten test folder
