+++ assets/scripts/palette/EnvironmentButton.test.js
@@ -2,11 +2,11 @@
 import React from 'react'
 import { screen } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { render } from '../../../../test/helpers/render'
-import EnvironmentButton from '../EnvironmentButton'
-import { toggleToolbox } from '../../store/slices/ui'
+import { render } from '../../../test/helpers/render'
+import { toggleToolbox } from '../store/slices/ui'
+import EnvironmentButton from './EnvironmentButton'
 
-jest.mock('../../store/slices/ui', () => ({
+jest.mock('../store/slices/ui', () => ({
   // We don't use these actions for anything, but they must return
   // a plain object or the dispatch() throws an error
   toggleToolbox: jest.fn((id) => ({ type: 'MOCK_ACTION' }))
