+++ assets/scripts/palette/PaletteTrashcan.test.js
@@ -1,8 +1,8 @@
 /* eslint-env jest */
 import React from 'react'
 import { screen } from '@testing-library/react'
-import { render } from '../../../../test/helpers/render'
-import PaletteTrashcan from '../PaletteTrashcan'
+import { render } from '../../../test/helpers/render'
+import PaletteTrashcan from './PaletteTrashcan'
 
 describe('PaletteTrashcan', () => {
   it('renders when visible', () => {
